--- src/main/java/com/acorn/api/com/acorn/api/WarehouseProjectApplication.java
@@ -1,13 +1,0 @@
-package com.acorn.api.com.acorn.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WarehouseProjectApplication {
-	
-	// 리펙토링 시작
-	public static void main(String[] args) {
-		SpringApplication.run(WarehouseProjectApplication.class, args);
-	}
-}
