+++ src/main/java/com/acorn/api/controller/owner/OwnerValidationController.java
@@ -1,9 +1,10 @@
 package com.acorn.api.controller.owner;
 
-import com.acorn.api.com.acorn.api.code.code.common.ApiValidationErrorCode;
-import com.acorn.api.com.acorn.api.code.code.common.ApiValidationSuccessCode;
-import com.acorn.api.com.acorn.api.code.response.ApiErrorResponse;
-import com.acorn.api.com.acorn.api.code.response.ApiSuccessResponse;
+
+import com.acorn.api.code.common.ApiValidationErrorCode;
+import com.acorn.api.code.common.ApiValidationSuccessCode;
+import com.acorn.api.code.response.ApiErrorResponse;
+import com.acorn.api.code.response.ApiSuccessResponse;
 import com.acorn.api.service.owner.OwnerService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
