+++ src/main/java/com/acorn/api/security/CustomLoginFailureHandler.java
@@ -1,8 +1,8 @@
 package com.acorn.api.security;
 
 
-import com.acorn.api.com.acorn.api.code.code.user.ApiUserErrorCode;
-import com.acorn.api.com.acorn.api.code.response.ApiErrorResponse;
+import com.acorn.api.code.user.ApiUserErrorCode;
+import com.acorn.api.code.response.ApiErrorResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
